Update animation.js
The async(0 in post message was causing an error so going the function() route

Also - i had a question to preview the animation will need a button to toggle visibility etc. 
but dont need it since only the preview folder has _a_ layers. Photopea only looks at this in the export gif
so other layers can remain as they are